+++ app/memo/(components)/memo-system-navigator/memo-system-navigator.tsx
@@ -1,9 +1,8 @@
 'use client'
 import React, { useCallback, useEffect, useState } from 'react'
-import { type Folder, isFolderHasMemo } from '../../(domain)/folder'
+import { type Folder, isFolderHasMemo, rebuildNewNameFolder } from '../../(domain)/folder'
 import { changeFolderName } from '../../(infra)/memo'
 import MemoFolderContextMenu, { type ContextMenuProps } from './memo-folder-context-menu'
-import { rebuildNewNameFolder } from './memo-folder-hooks'
 import { FolderAndMemo } from './FolderAndMemoStructure'
 import { type Tab } from '../../../(layout)/(domain)/tab'
 import { useTabBarAndRouter } from '../../../(layout)/(usecase)/tab-usecases'
